Meni

+++ BFS.h
@@ -41,7 +41,7 @@ public:
         }
     }
 
-    void WriteDirection(int decideDirection, double totalCost) {
+    void WriteDirection(int decideDirection, int totalCost) {
         string extra;
         switch (decideDirection) {
             case 1:
